refactor: extract out common inline edit input
