+++ apps/journeys-admin/src/components/Editor/Canvas/InlineEditWrapper/InlineEditInput/InlineEditInput.tsx
@@ -1,0 +1,23 @@
+import { styled, SimplePaletteColorOptions } from '@mui/material/styles'
+import InputBase, { InputBaseProps } from '@mui/material/InputBase'
+import { adminTheme } from '../../../../ThemeProvider/admin/theme'
+
+interface StyledInputProps extends InputBaseProps {}
+
+const adminPrimaryColor = adminTheme.palette
+  .primary as SimplePaletteColorOptions
+
+export const InlineEditInput = styled(InputBase)<StyledInputProps>(() => ({
+  '& .MuiInputBase-input': {
+    textAlign: 'inherit',
+    textTransform: 'inherit'
+  },
+  color: 'inherit',
+  fontSize: 'inherit',
+  fontWeight: 'inherit',
+  fontFamily: 'inherit',
+  lineHeight: 'inherit',
+  letterSpacing: 'inherit',
+  padding: '0px',
+  caretColor: adminPrimaryColor.main
+}))
