+++ apps/journeys-admin/src/components/Editor/Canvas/InlineEditWrapper/InlineEditInput/index.ts
@@ -1,0 +1,1 @@
+export { InlineEditInput } from './InlineEditInput'
